requirements.txt
Django==3.2.16
django-bootstrap5==22.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
python-dotenv==1.0.0